Factorize dump implementation out of annotation
