+++ cytomine/utilities/__init__.py
@@ -22,3 +22,4 @@ from .geometry import *
 from .logging import *
 from .reader import *
 from .wholeslide import *
+from .download_util import download_annotation_crops
